+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 - Nested errors
 - tests: fix empty error names in `.to.be.instanceof(...)`
 - Implemented plugins support
+- odbm: Implemented isNew()
+- odbm: Support for custom ids
 
 ## 0.6.21